Moving command to core
