+++ src/bin/vip-dev-env-import-sql.js
@@ -17,9 +17,8 @@ import path from 'path';
  */
 import command from 'lib/cli/command';
 import { getEnvironmentName, handleCLIException } from 'lib/dev-environment/dev-environment-cli';
-import { getEnvironmentPath } from 'lib/dev-environment/dev-environment-core';
-import { exec } from 'lib/dev-environment/dev-environment-core';
-import { searchAndReplace } from 'lib/search-and-replace';
+import { exec, resolveImportPath } from 'lib/dev-environment/dev-environment-core';
+
 import { DEV_ENVIRONMENT_FULL_COMMAND } from 'lib/constants/dev-environment';
 
 // Command examples
@@ -43,45 +42,13 @@ command( {
 	.examples( examples )
 	.argv( process.argv, async ( unmatchedArgs: string[], opt ) => {
 		const [ fileName ] = unmatchedArgs;
-
 		const { searchReplace, inPlace } = opt;
 		const slug = getEnvironmentName( opt );
 
 		try {
-			let resolvedPath = path.resolve( fileName );
-
-			if ( ! fs.existsSync( resolvedPath ) ) {
-				throw new Error( 'The provided file does not exist or it is not valid (see "--help" for examples)' );
-			}
-
-			// Run Search and Replace if the --search-replace flag was provided
-			if ( searchReplace && searchReplace.length ) {
-				const { outputFileName } = await searchAndReplace( resolvedPath, searchReplace, {
-					isImport: true,
-					output: true,
-					inPlace,
-				} );
-
-				if ( typeof outputFileName !== 'string' ) {
-					throw new Error( 'Unable to determine location of the intermediate search & replace file.' );
-				}
-
-				const environmentPath = getEnvironmentPath( slug );
-				const baseName = path.basename( outputFileName );
-
-				resolvedPath = path.join( environmentPath, baseName );
-				fs.renameSync( outputFileName, resolvedPath );
-			}
-
-			const dockerPath = resolvedPath.replace( os.homedir(), '/user' );
-			const arg = [ 'wp', 'db', 'import', dockerPath ];
-
+			const resolvedPath = await resolveImportPath( slug, fileName, searchReplace, inPlace );
+			const arg = [ 'wp', 'db', 'import', resolvedPath ];
 			await exec( slug, arg );
-
-			// Removing search and replace temp SQL file
-			if ( searchReplace && searchReplace.length && ! inPlace ) {
-				fs.unlinkSync( resolvedPath );
-			}
 		} catch ( e ) {
 			handleCLIException( e );
 		}
